RavenDB-23460 Allow deleting revisions only for DatabaseAdmin
